Initial implementation of recovery from checkpoint. (#411)
* tmp

* Support header only in consensus.

* Merge branch 'master' into header_only

* Handle ignore_body.

* nit.

* Remove on_new_block_construction_only.

* Remove construct_pivot_info().

* Merge branch 'master' into fullnode

* Merge branch 'header_only' into fullnode

* Pass tests.

* Add fullnode test.

* Delay clone to tx pool.

* Merge branch 'master' into header_only

* Merge branch 'header_only' into fullnode

* Remove unused code.

* Merge branch 'master' into fullnode

* fmt.

* Fix some comments.

* Add FullNode.

* Bug fix.

* Bug fix again.

+++ client/src/rpc/traits/cfx/test.rs
@@ -21,7 +21,7 @@ pub trait TestRpc {
     fn get_best_block_hash(&self) -> RpcResult<H256>;
 
     #[rpc(name = "getblockcount")]
-    fn get_block_count(&self) -> RpcResult<usize>;
+    fn get_block_count(&self) -> RpcResult<u64>;
 
     #[rpc(name = "getgoodput")]
     fn get_goodput(&self) -> RpcResult<isize>;
